Move ValueChecker.h and ConstraintsChecker.h to src/app/tests/suites/include
